refac(BottomToolbar):  move TagPopper into BottomToolbar TagPopper was previously in the AddTagButton as a sibling component.  It made more sense for TagPopper to be in the ImageBottomToolbar.  This makes AddTagButton a unit that will be easier to test.
