+++ Frontend/src/ImageView/BottomToolbar/ImageBottomToolbar.js
@@ -1,26 +1,12 @@
-import React, { useContext } from 'react';
+import React, { useContext, useRef, useState } from 'react';
 import { Toolbar } from "@mui/material";
 
 import Tag from './Tag';
 import AddTagButton from './AddTagButton';
+import TagPopper from './TagPopper';
 import ExcessTagsChip from './ExcessTagsChip';
 import ImageDataContext from '../ImageDataContext';
 
-
-/**
-  * Creates the tags to be displayed on the toolbar.
-  *
-  * @param {array} imageTags Array of tag IDs associated with this image.
-  * @param {number} imageId Unique identifier of the image with these tags.
-  * @returns A Tag component for each tag associated with this image.
-  */
-const generateTags = (imageTags, imageId) => {
-  return(
-    imageTags.map((imageTag) => (
-      <Tag imageTag={imageTag} imageId={imageId} key={imageTag} />
-    ))
-  );
-};
 
 /**
   * Formats the tags to be displayed on the toolbar.
@@ -35,16 +21,23 @@ const formatTags = (imageTags, imageId) => {
     return (<Tag imageTag={0} key={0} />); // display "add tag" placeholder
 
   } else if (imageTags.length < 3) {
-    return generateTags(imageTags, imageId); // display normally
+    // display normally
+    return (
+      imageTags.map((imageTag) => (
+        <Tag imageTag={imageTag} imageId={imageId} key={imageTag} />
+      ))
+    );
 
   } else {
     // display condensed, with "+excess" chip
     return (
       <>
-        {generateTags(imageTags, imageId).slice(0, 2) /* show only two tags */}
+        {imageTags.map((imageTag) => (
+          <Tag imageTag={imageTag} imageId={imageId} key={imageTag} />
+        )).slice(0, 2) /* show only two tags */}
         <ExcessTagsChip labelText={(imageTags.length - 2).toString()} />
       </>
-    )
+    );
   };
 };
 
@@ -62,10 +55,31 @@ function ImageBottomToolbar(props) {
   const imageId = imageData['id'];
   const imageTags = imageData['imageTags'];
 
+  // See https://mui.com/material-ui/react-popper/ for details on the purpose of anchorEl.
+  const [anchorEl, setAnchorEl] = useState(null);
+  
+  // Sets AddTagPopper open or closed.
+  const [open, setOpen] = useState(false);
+  
+  // Used by AddTagPopper for click-away handling.
+  const buttonRef = useRef(null);
+
   return(
     <Toolbar sx={[props.toolbarStyles, { bottom: 6.5 }]}>
       {formatTags(imageTags, imageId)} {/* Returns Tag components for image */}
-      <AddTagButton style={{ marginLeft: "auto" }}/>
+      <AddTagButton
+        style={{ marginLeft: "auto" }}
+        buttonRef={buttonRef}
+        open={open}
+        setOpen={setOpen}
+        anchorEl={anchorEl}
+        setAnchorEl={setAnchorEl}/>
+      <TagPopper
+        open={open}
+        setOpen={setOpen}
+        anchorEl={anchorEl}
+        setAnchorEl={setAnchorEl}
+        buttonRef={buttonRef}/>
     </Toolbar>
   );
 };
